Further shrinking of the Dockerfile
So, after some discussion, it seems best to just carry one using
a basic Ubuntu image (ussing focal here), and putting all the
jupyter stacks stuff in.

I have also updated the package list to be more sensible, and applied
some of the trics that @MarcYin identified to clear further packages.
The final size is ~590Mb, which is less than a CD (if anyone still
remembers them).

The docker file is also split in two, so it's easy to just add
packages on top of the basic image.

I have not installed latex, but it is a possibility.